Add MicroFrontend2 and refactor Router class
This commit includes the addition of the microfrontend for the About page (MicroFrontend2). It also includes a major refactor of the 'Router' class for better control of the loading process of microfrontends, by introducing the 'MicroFrontends' enum. Furthermore, scripts for "start" and "build" for both the whole application and specifically the "about" module have been added to the package.json file.