Update CI/CD workflow and add code analysis
Updated the CI/CD workflow file (ci-cd.yml) with a more explicit name. Inclusion of a new 'analyze' job has been done to carry out CodeQL analysis on the codebase. This job runs a series of steps including checkout, initialization of CodeQL, autobuild and analysis. Also, refactored the process of 'push' under 'on' to make it more concise.